source/maya_aceclient/tests/mock_ace_server/requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
numpy==1.23.5
protobuf==5.26.1
